requirements.txt
streamlit 
transformers 
torch
sentencepiece
SpeechRecognition 
pydub
streamlit-audiorecorder
tensorflow
tensorflow-hub 
pillow 
matplotlib
numpy